requirements.txt
python-telegram-bot==21.4
requests==2.28.1
pandas==2.2.2
openpyxl==3.1.5
